+++ api/routes/user.js
@@ -1,117 +1,12 @@
 const express = require('express');
 const router = express.Router();
-const bcrypt = require('bcrypt');
-const mongoose = require('mongoose');
-const jwt = require('jsonwebtoken');
+const usersController = require('../controllers/users');
+const checkAuth = require('../middleware/check-auth');
 
-const User = require('../models/user');
+router.post('/signup', usersController.users_signup_user);
 
-router.post('/signup', (req, res, next) => {
-    User.find({email: req.body.email})
-        .exec()
-        .then(user => {
-            if (user.length >= 1) {
-                return res.status(409).json({
-                    message: 'Mail exists'
-                });
-            } else {
-                bcrypt.hash(req.body.password, 10, (err, hashedPassword) => {
-                    if (err) {
-                        console.log(err);
-                        return res.status(500).json({
-                            error: err
-                        });
-                    } else {
-                        const user = new User({
-                            _id: mongoose.Types.ObjectId(),
-                            email: req.body.email,
-                            password: hashedPassword
-                        }).save()
-                        .then(result => {
-                            console.log(result);
-                            res.status(201).json({
-                                message: 'User created'
-                            });
-                        })
-                        .catch(error => {
-                            console.log(error);
-                            res.status(500).json({
-                                error: error
-                            });
-                        });
-                    }
-                });
-            }
-        })
-        .catch(error => {
-            console.log(error);
-            res.status(500).json({
-                error: error
-            });
-        });
-});
+router.post('/login', usersController.users_login_user);
 
-router.post('/login', (req, res, next) => {
-    User.find( {email: req.body.email} )
-        .exec()
-        .then(user => {
-            if (user.length < 1) {
-                return res.status(401).json({
-                    message: 'Auth faild'
-                });
-            }
-            bcrypt.compare(req.body.password, user[0].password, (error, result) => {
-                if (error) {
-                    console.log(error);
-                    res.status(500).json({
-                        error: error
-                    });
-                }
-                const token = jwt.sign(
-                    {
-                        email: user[0].email,
-                        userId: user[0]._id
-                    },
-                    process.env.JWT_KEY,
-                    {
-                        expiresIn: '1h'
-                    }
-                );
-                if (result) {
-                    return res.status(201).json({
-                        message: 'Auth successfully',
-                        token: token
-                    });
-                }
-
-                res.status(401).json({
-                    message: 'Auth faild'
-                });
-            });
-        })
-        .catch(error => {
-            console.log(error);
-            res.status(500).json({
-                error: error
-            });
-        });
-});
-
-router.delete('/:userId', (req, res, next) => {
-    const userId = req.params.userId;
-    User.remove( {_id: userId} )
-        .exec()
-        .then(result => {
-            res.status(200).json({
-                message: 'User deleted'
-            });
-        })
-        .catch(error => {
-            console.log(error);
-            res.status(500).json({
-                error: error
-            });
-        });
-})
+router.delete('/:userId', checkAuth, usersController.users_delete_user);
 
 module.exports = router;
